apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cover-letter-ai-api"
version = "0.1.0"
description = "FastAPI backend for cover-letter-ai"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "python-multipart>=0.0.9",
  "pydantic>=2.7",
  "pydantic-settings>=2.2",
  "openai>=1.68.0",
  "firecrawl-py>=2.0.0",
  "docxtpl>=0.19.0",
  "pypdf>=5.0.0",
  "python-dotenv>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]


